+++ src/helpers/Localization.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * SEOmatic plugin for Craft CMS 3.x
+ *
+ * A turnkey SEO implementation for Craft CMS that is comprehensive, powerful,
+ * and flexible
+ *
+ * @link      https://nystudio107.com
+ * @copyright Copyright (c) 2019 nystudio107
+ */
+
+namespace nystudio107\seomatic\helpers;
+
+/**
+ * @author    nystudio107
+ * @package   Seomatic
+ * @since     3.2.30
+ */
+class Localization
+{
+    // Public Static Properties
+    // =========================================================================
+
+    // Public Static Methods
+    // =========================================================================
+
+    /**
+     * Normalizes a language into the correct format (e.g. `de_DE`).
+     *
+     * @param string $language
+     *
+     * @return string
+     */
+    public static function normalizeOgLocaleLanguage(string $language): string
+    {
+        $language = str_replace('-', '_', $language);
+        if (strlen($language) === 2) {
+            $language = strtolower($language).'_'.strtoupper($language);
+        }
+
+        return $language;
+    }
+}